restructure files and improve imports
